07-MCP/requirements.txt
fastmcp
python-dotenv
pandas
langchain-groq
langchain-openai
langchain-community
faiss-cpu
openpyxl
mcp